Add files via upload
This program you can input your order burger, drinks and you can input also your Cash on Hand. It will display in richtextbox the total price of every item and also the total price of all item did you choose. It also display the Cash on Hand and the change.